+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.9-beta.0](https://prifina-admin/prifina/app-desktop/compare/@prifina-apps/utils@0.0.8-beta.0...@prifina-apps/utils@0.0.9-beta.0) (2022-04-30)
+
+**Note:** Version bump only for package @prifina-apps/utils
+
+
+
+
+
+## [0.0.8-beta.0](https://prifina-admin/prifina/app-desktop/compare/@prifina-apps/utils@0.0.3-beta.0...@prifina-apps/utils@0.0.8-beta.0) (2022-04-08)
+
+**Note:** Version bump only for package @prifina-apps/utils
+
+
+
+
+
 ## [0.0.4-beta.0](https://prifina-admin/prifina/app-desktop/compare/@prifina-apps/utils@0.0.3-beta.0...@prifina-apps/utils@0.0.4-beta.0) (2022-03-07)
 
 **Note:** Version bump only for package @prifina-apps/utils